Moved mysnackbarContent
Moved MySnackbarContent out of app.jsx just to clean up this component
